Chore: Migrate selector/login to TS (#3731)
* migrate selector/login to TS

* Fix lint errors

* set aliases for returns
